# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/availability": {
      "get": {
        "description": "An empty `slots` array is a valid, non-error response (fully booked or a closed day). Redis-cached ~60s.",
        "operationId": "AvailabilityController_getAvailability",
        "parameters": [
          {
            "name": "serviceId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "Local calendar date in the clinic timezone",
            "schema": {
              "example": "2026-03-05",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponseDto"
                }
              }
            }
          }
        },
        "summary": "Bookable slots for a service on a local calendar date",
        "tags": [
          "availability"
        ]
      }
    },
    "/api/v1/admin/availability-blocks": {
      "get": {
        "operationId": "AvailabilityBlocksController_list",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-03-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-04-01T00:00:00.000Z",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailabilityBlockDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List availability blocks overlapping a range",
        "tags": [
          "admin/availability-blocks"
        ]
      },
      "post": {
        "description": "Overlapping an existing booking always succeeds — the booking is never cancelled, only reported back in affectedAppointments.",
        "operationId": "AvailabilityBlocksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAvailabilityBlockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityBlockWithWarningsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an availability block",
        "tags": [
          "admin/availability-blocks"
        ]
      }
    },
    "/api/v1/admin/availability-blocks/{id}": {
      "patch": {
        "operationId": "AvailabilityBlocksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAvailabilityBlockDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityBlockWithWarningsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an availability block",
        "tags": [
          "admin/availability-blocks"
        ]
      },
      "delete": {
        "operationId": "AvailabilityBlocksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an availability block",
        "tags": [
          "admin/availability-blocks"
        ]
      }
    },
    "/api/v1/appointments": {
      "post": {
        "operationId": "AppointmentsController_book",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "SLOT_UNAVAILABLE"
          },
          "422": {
            "description": "SLOT_OUTSIDE_AVAILABILITY | CONSENT_REQUIRED | COUPON_*"
          }
        },
        "summary": "Guest booking",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/appointments/reschedule/{token}": {
      "get": {
        "description": "Never returns the full patient record — name and phone only. Single-use: a second call is RESCHEDULE_TOKEN_INVALID.",
        "operationId": "AppointmentsController_resolveReschedule",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReschedulePrefillDto"
                }
              }
            }
          },
          "404": {
            "description": "RESCHEDULE_TOKEN_INVALID"
          }
        },
        "summary": "Resolve a signed rebooking link to prefill data",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/admin/appointments": {
      "get": {
        "operationId": "AdminAppointmentsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "booked",
                "completed",
                "cancelled",
                "no_show"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-03-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-04-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Matches patient name or phone",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAdminAppointmentsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List appointments, filtered and paginated",
        "tags": [
          "admin/appointments"
        ]
      },
      "post": {
        "description": "Same slot rules as the public booking endpoint.",
        "operationId": "AdminAppointmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAppointmentDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manual / walk-in booking",
        "tags": [
          "admin/appointments"
        ]
      }
    },
    "/api/v1/admin/appointments/{id}": {
      "get": {
        "operationId": "AdminAppointmentsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAppointmentDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Appointment detail",
        "tags": [
          "admin/appointments"
        ]
      },
      "patch": {
        "operationId": "AdminAppointmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAppointmentDetailDto"
                }
              }
            }
          },
          "409": {
            "description": "SLOT_UNAVAILABLE"
          },
          "422": {
            "description": "SLOT_OUTSIDE_AVAILABILITY"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reschedule or change the status of an appointment",
        "tags": [
          "admin/appointments"
        ]
      }
    },
    "/api/v1/admin/appointments/{id}/cancel": {
      "post": {
        "operationId": "AdminAppointmentsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAppointmentDetailDto"
                }
              }
            }
          },
          "422": {
            "description": "APPOINTMENT_NOT_CANCELLABLE"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel an appointment",
        "tags": [
          "admin/appointments"
        ]
      }
    }
  },
  "info": {
    "title": "Physio Clinic API",
    "description": "Booking, patient records and clinic administration for the physiotherapy clinic. This specification is the contract the website and CRM generate their types from.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "CRM access token from POST /auth/google or /auth/login."
      }
    },
    "schemas": {
...
      "AvailabilityResponseDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "example": "2026-03-05"
          },
          "timezone": {
            "type": "string",
            "example": "Asia/Kolkata"
          },
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "durationMinutes": {
            "type": "number",
            "example": 45
          },
          "slots": {
            "description": "Bookable slots. An empty array is a valid, non-error response.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotDto"
            }
          }
        },
        "required": [
          "date",
          "timezone",
          "serviceId",
          "durationMinutes",
          "slots"
        ]
      },
      "AvailabilityBlockDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "string",
            "example": "2026-03-05T04:30:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2026-03-05T09:30:00.000Z"
          },
          "reason": {
            "type": "object",
            "nullable": true,
            "example": "Personal leave"
          },
          "recurringRule": {
            "type": "object",
            "nullable": true,
            "example": "FREQ=WEEKLY;COUNT=6"
          }
        },
        "required": [
          "id",
          "startsAt",
          "endsAt",
          "reason",
          "recurringRule"
        ]
      },
      "CreateAvailabilityBlockDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "example": "2026-03-05T04:30:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2026-03-05T09:30:00.000Z"
          },
          "reason": {
            "type": "string",
            "example": "Personal leave"
          },
          "recurringRule": {
            "type": "string",
            "example": "FREQ=WEEKLY;COUNT=6",
            "description": "RRULE string, anchored to startsAt as DTSTART."
          }
        },
        "required": [
          "startsAt",
          "endsAt"
        ]
      },
...
      "AvailabilityBlockWithWarningsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "string",
            "example": "2026-03-05T04:30:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2026-03-05T09:30:00.000Z"
          },
          "reason": {
            "type": "object",
            "nullable": true,
            "example": "Personal leave"
          },
          "recurringRule": {
            "type": "object",
            "nullable": true,
            "example": "FREQ=WEEKLY;COUNT=6"
          },
          "affectedAppointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AffectedAppointmentDto"
            }
          }
        },
        "required": [
          "id",
          "startsAt",
          "endsAt",
          "reason",
          "recurringRule",
          "affectedAppointments"
        ]
      },
      "UpdateAvailabilityBlockDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "example": "2026-03-05T04:30:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2026-03-05T09:30:00.000Z"
          },
          "reason": {
            "type": "string",
            "example": "Personal leave"
          },
          "recurringRule": {
            "type": "string",
            "example": "FREQ=WEEKLY;COUNT=6"
          }
        }
      },
...
      "CreateAppointmentDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "string",
            "example": "2026-03-05T04:30:00.000Z"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientDetailsDto"
          },
          "reasonForVisit": {
            "type": "string",
            "enum": [
              "acute_injury",
              "persistent_pain",
              "post_surgical",
              "chronic_condition",
              "sports_performance",
              "general_assessment"
            ]
          },
          "couponCode": {
            "type": "object",
            "example": "WELCOME10",
            "nullable": true
          },
          "paymentPreference": {
            "type": "string",
            "enum": [
              "online",
              "clinic"
            ]
          },
          "consentGiven": {
            "type": "boolean",
            "example": true
          },
          "captchaToken": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "serviceId",
          "startsAt",
          "patient",
          "reasonForVisit",
          "paymentPreference",
          "consentGiven"
        ]
      },
...
      "AppointmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reference": {
            "type": "string",
            "example": "PH-2026-0001"
          },
          "startsAt": {
            "type": "string",
            "example": "2026-03-05T04:30:00.000Z"
          },
          "endsAt": {
            "type": "string",
            "example": "2026-03-05T05:15:00.000Z"
          },
          "status": {
            "type": "string",
            "example": "booked"
          },
          "service": {
            "$ref": "#/components/schemas/AppointmentServiceDto"
          },
          "patient": {
            "$ref": "#/components/schemas/AppointmentPatientDto"
          },
          "pricing": {
            "$ref": "#/components/schemas/AppointmentPricingDto"
          },
          "paymentStatus": {
            "type": "string",
            "example": "pending"
          }
        },
        "required": [
          "id",
          "reference",
          "startsAt",
          "endsAt",
          "status",
          "service",
          "patient",
          "pricing",
          "paymentStatus"
        ]
      },
...
      "ReschedulePrefillDto": {
        "type": "object",
        "properties": {
          "patient": {
            "$ref": "#/components/schemas/ReschedulePatientDto"
          },
          "suggestedServiceId": {
            "type": "string",
            "format": "uuid"
          },
          "previousAppointmentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "patient",
          "suggestedServiceId",
          "previousAppointmentId"
        ]
      },
...
      "PaginatedAdminAppointmentsDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminAppointmentListItemDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "AdminAppointmentDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reference": {
            "type": "string",
            "example": "PH-2026-0001"
          },
          "startsAt": {
            "type": "string"
          },
          "endsAt": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "service": {
            "$ref": "#/components/schemas/AdminAppointmentServiceDto"
          },
          "patient": {
            "$ref": "#/components/schemas/AdminAppointmentPatientDto"
          },
          "priceCharged": {
            "type": "string",
            "example": "900.00"
          },
          "paymentStatus": {
            "type": "string"
          },
          "bookingSource": {
            "type": "string"
          },
          "originalPrice": {
            "type": "string",
            "example": "1000.00"
          },
          "discountAmount": {
            "type": "string",
            "example": "100.00"
          },
          "couponCode": {
            "type": "object",
            "nullable": true
          },
          "reasonForVisit": {
            "type": "string"
          },
          "paymentPreference": {
            "type": "string"
          },
          "cancellationReason": {
            "type": "object",
            "nullable": true
          },
          "cancelledBy": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AdminAppointmentCancelledByDto"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "startsAt",
          "endsAt",
          "status",
          "service",
          "patient",
          "priceCharged",
          "paymentStatus",
          "bookingSource",
          "originalPrice",
          "discountAmount",
          "couponCode",
          "reasonForVisit",
          "paymentPreference",
          "cancellationReason",
          "cancelledBy",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateAdminAppointmentDto": {
        "type": "object",
        "properties": {
          "serviceId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "string",
            "example": "2026-03-05T04:30:00.000Z"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientDetailsDto"
          },
          "reasonForVisit": {
            "type": "string",
            "enum": [
              "acute_injury",
              "persistent_pain",
              "post_surgical",
              "chronic_condition",
              "sports_performance",
              "general_assessment"
            ]
          },
          "couponCode": {
            "type": "object",
            "example": "WELCOME10",
            "nullable": true
          },
          "paymentPreference": {
            "type": "string",
            "enum": [
              "online",
              "clinic"
            ]
          },
          "consentGiven": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "serviceId",
          "startsAt",
          "patient",
          "reasonForVisit",
          "paymentPreference",
          "consentGiven"
        ]
      },
      "UpdateAppointmentDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "example": "2026-03-06T04:30:00.000Z"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "no_show"
            ]
          }
        }
      },
      "CancelAppointmentDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Doctor unwell, rescheduling on request"
          }
        },
        "required": [
          "reason"
        ]
      }
    }
  }
}